requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langgraph
langchain-community
yfinance
requests
python-dotenv
pydantic
httpx
pandas
numpy
ta
slowapi
bcrypt
python-jose[cryptography]
aiosqlite
pydantic[email]